Time: 1 ms (99.57%) | Memory: 56.6 MB (97.10%) - LeetSync

+++ 53-maximum-subarray/maximum-subarray.java
@@ -1,7 +1,7 @@
 class Solution {
     public int maxSubArray(int[] nums) {
-        long max=Long.MIN_VALUE;
-        long sum=0;
+        int max=Integer.MIN_VALUE;
+        int sum=0;
         for(int i=0;i<nums.length;i++){
            // if(nums[i]>max){
                  sum+=nums[i];
@@ -15,6 +15,6 @@ class Solution {
 
 
         }
-        return (int)max;
+        return max;
     }
 }
